requirements.txt
django==3.1.5
djangorestframework==3.12.2
django-filter==2.4.0
apscheduler==3.7.0
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1


